Refactor furioso y reacomode
