+++ .eslintrc.json
@@ -20,6 +20,7 @@
     "semi": ["error", "never"],
     "class-methods-use-this": "off",
     "comma-dangle": "off",
-    "eol-last": "off"
+    "eol-last": "off",
+    "no-plusplus": "off"
   }
 }
