+++ src/norms/criterias/ItemCriteria/ItemPrescribeCriteria/ItemPrescribedCriteria.js
@@ -1,0 +1,9 @@
+const { ItemCriteria } = require('../itemCriteria')
+
+class ItemPresribedCriteria extends ItemCriteria {
+  getEntity() {
+    return 'ITEM_PRESCRIBED'
+  }
+}
+
+module.exports = { ItemPresribedCriteria }
