+++ src/norms/metadata/rulesMetadata.js
@@ -1,0 +1,11 @@
+const { Rule } = require('../rules/rule')
+const { loadPredicate } = require('./metadata')
+
+console.log(loadPredicate)
+
+const RULE_METADATA = json => new Rule().initialize({
+  ...json,
+  predicate: loadPredicate(json)
+})
+
+module.exports = { RULE_METADATA }
