+++ src/norms/norm.js
@@ -1,0 +1,6 @@
+const { generateJsonFromRule, generateRuleFromJson } = require('./metadata/metadata')
+
+module.exports = {
+  generateJsonFromRule,
+  generateRuleFromJson
+}
