+++ src/norms/predicates/predicate.js
@@ -3,9 +3,6 @@
 // Esta es la clase principal de la cual heredan 4 mas
 // Es un command. Es decir, primero se carga con valores (initialize)
 // y despues se manda a ejecutar (satisfies)
-// Posiblemente esta clase tenga que tener un metodo toJson que devuelva
-// la representacion de todo el arbol que se forme en formato json y tambien
-// un from json. Pero hay que pensarlo bien
 class Predicate {
   // Cada una de las subclases sabe que tipo de modelo espera
   // por lo que se delega la validacion e inicializacion a las mismas
@@ -38,6 +35,7 @@ class Predicate {
     throw new Error('Template method. Please override!')
   }
 
+  // Se obtiene la representacion en json del arbol de predicados
   toJson() {
     throw new Error('Template method. Please override!')
   }
